refactor: Move ai.ts to services/ directory
Consolidates AI service into services/ directory where image-generation
already lives. Updates all agent imports from utils/ai to services/ai.

Closes #34
