+++ src/core/agents/visuals.ts
@@ -1,4 +1,4 @@
-import { generateObject, streamObjectWithProgress } from '../utils/ai';
+import { generateObject, streamObjectWithProgress } from '../services/ai';
 import { z } from 'zod';
 import {
   VisualDirectionSchema,
